+++ src/codepages.rs
@@ -1,0 +1,35 @@
+//! Support for working with different codepages / encodings.
+
+use crate::error::DecodeError;
+use std::borrow::Cow;
+
+/// Trait for reading strings from the database files.
+///
+/// If the `yore` feature isn't on, this is implemented only be [`UnicodeLossy`].
+///
+/// If the `yore` feature is on, this is implemented by all [`yore::CodePage`].
+///
+/// Note: This trait might be extended with an `encode` function in the future.
+pub trait Encoding {
+    /// Decode encoding into UTF-8 string. If codepoints can't be represented, an error is returned.
+    fn decode<'a>(&self, bytes: &'a [u8]) -> Result<Cow<'a, str>, DecodeError>;
+}
+
+pub struct UnicodeLossy;
+
+/// Tries to decode as Unicode, replaces unknown codepoints with the replacement character.
+impl Encoding for UnicodeLossy {
+    fn decode<'a>(&self, bytes: &'a [u8]) -> Result<Cow<'a, str>, DecodeError> {
+        Ok(String::from_utf8_lossy(bytes))
+    }
+}
+
+#[cfg(feature = "yore")]
+impl<T> Encoding for T
+where
+    T: yore::CodePage,
+{
+    fn decode<'a>(&self, bytes: &'a [u8]) -> Result<Cow<'a, str>, DecodeError> {
+        self.decode(bytes).map_err(Into::into)
+    }
+}