+++ src/de.rs
@@ -5,10 +5,11 @@ use serde::de::{DeserializeOwned, DeserializeSeed, IntoDeserializer, SeqAccess, 
 use serde::Deserializer;
 
 use crate::{
-    ErrorKind, FieldConversionError, FieldIOError, FieldIterator, FieldValue, ReadableRecord,
+    Encoding, ErrorKind, FieldConversionError, FieldIOError, FieldIterator, FieldValue,
+    ReadableRecord,
 };
 
-impl<'de, 'a, 'f, R: Read + Seek> SeqAccess<'de> for &mut FieldIterator<'a, R> {
+impl<'de, 'a, 'f, R: Read + Seek, E: Encoding> SeqAccess<'de> for &mut FieldIterator<'a, R, E> {
     type Error = FieldIOError;
 
     fn next_element_seed<T>(
@@ -27,7 +28,7 @@ impl<'de, 'a, 'f, R: Read + Seek> SeqAccess<'de> for &mut FieldIterator<'a, R> {
 }
 
 //TODO maybe we can deserialize numbers other than f32 & f64 by converting using TryFrom
-impl<'de, 'a, 'f, T: Read + Seek> Deserializer<'de> for &mut FieldIterator<'a, T> {
+impl<'de, 'a, 'f, T: Read + Seek, E: Encoding> Deserializer<'de> for &mut FieldIterator<'a, T, E> {
     type Error = FieldIOError;
 
     fn deserialize_any<V>(self, _visitor: V) -> Result<<V as Visitor<'de>>::Value, Self::Error>
@@ -324,9 +325,10 @@ impl<'de, 'a, 'f, T: Read + Seek> Deserializer<'de> for &mut FieldIterator<'a, T
 }
 
 impl<S: DeserializeOwned> ReadableRecord for S {
-    fn read_using<T>(field_iterator: &mut FieldIterator<T>) -> Result<Self, FieldIOError>
+    fn read_using<T, E>(field_iterator: &mut FieldIterator<T, E>) -> Result<Self, FieldIOError>
     where
         T: Read + Seek,
+        E: Encoding,
     {
         S::deserialize(field_iterator)
     }