+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next
     - Replaced `chrono` with `time` v0.3
+    - Added support for reading non-unicode database files via custom encodings.
+      The optional `yore` feature/crate can be used for support for basic codepages.
 
 # 0.2.3
     - Added `impl std::error::Error for dbase::Error`